rearranged code into subpackages; fix concurrency
