+++ model/init.go
@@ -1,0 +1,10 @@
+package model
+
+import (
+	"github.com/charmbracelet/bubbles/textinput"
+	tea "github.com/charmbracelet/bubbletea"
+)
+
+func (m model) Init() tea.Cmd {
+	return textinput.Blink
+}
